cleaned up manager and password services structure. fixed unit tests
