+++ Backend/ParkingMaster.Services/ParkingMaster.Services.csproj
@@ -49,7 +49,6 @@
     <Compile Include="Services\IPasswordService.cs" />
     <Compile Include="Services\IUserManagementService.cs" />
     <Compile Include="Services\PasswordService.cs" />
-    <Compile Include="Services\PasswordValidationManager.cs" />
     <Compile Include="Services\PwnedPasswords.cs" />
     <Compile Include="Services\UserManagementService.cs" />
   </ItemGroup>
